+++ src/actions.ts
@@ -1,15 +1,15 @@
-import clipboardy from "clipboardy";
-import open from "open";
-import readline from "readline";
-import { ClickUp } from "./clickup";
-import { GitLab } from "./gitlab";
-import { NormalizedChecklist } from "./models/models";
+import clipboardy from 'clipboardy';
+import open from 'open';
+import readline from 'readline';
+import { ClickUp } from './clickup';
+import { GitLab } from './gitlab';
+import { NormalizedChecklist } from './models/models';
 import {
   getClickUpTaskIdFromGitLabIssue,
   getGitLabProjectConfigById,
   normalizeClickUpChecklist,
   normalizeGitLabIssueChecklist,
-} from "./utils";
+} from './utils';
 
 export function getSyncChecklistActions(
   oldClickUpChecklist: NormalizedChecklist,
@@ -48,17 +48,21 @@ export function getSyncChecklistActions(
 export async function syncChecklist(
   gitLabProjectId: string,
   issueNumber: string,
-  openIssuePage?: boolean
+  openPage?: boolean
 ) {
   const gitLab = new GitLab(gitLabProjectId);
   const issue = await gitLab.getIssue(issueNumber);
-  if (openIssuePage) {
+  if (openPage) {
     open(issue.web_url);
+    const result = issue.description.match(/https:\/\/app.clickup.com\/t\/\w+/);
+    if (result) {
+      open(result[0]);
+    }
   }
   const clickUpTaskId = getClickUpTaskIdFromGitLabIssue(issue);
   if (clickUpTaskId) {
     const gitLabChecklistText = issue.description
-      .replace(/https:\/\/app.clickup.com\/t\/\w+/g, "")
+      .replace(/https:\/\/app.clickup.com\/t\/\w+/g, '')
       .trim();
     const gitLabNormalizedChecklist = normalizeGitLabIssueChecklist(
       gitLabChecklistText
@@ -66,8 +70,8 @@ export async function syncChecklist(
     const clickUp = new ClickUp(clickUpTaskId);
     const clickUpTask = await clickUp.getTask();
     const clickUpChecklistTitle = `GitLab synced checklist [${gitLabProjectId.replace(
-      "%2F",
-      "/"
+      '%2F',
+      '/'
     )}]`;
     let clickUpChecklist = clickUpTask.checklists.find(
       (c: any) => c.name === clickUpChecklistTitle
@@ -115,19 +119,19 @@ export async function syncChecklist(
     const status = Object.entries(actions)
       .map(([action, items]) => {
         const s = items.length.toString();
-        const n = items.length === 1 ? "item" : "items";
+        const n = items.length === 1 ? 'item' : 'items';
         return `${s} ${n} ${action}d`;
       })
-      .join(", ");
+      .join(', ');
     const fullCompleteMessage = gitLabNormalizedChecklist.every(
       (item) => item.checked
     )
-      ? "(Completed)"
-      : "";
+      ? '(Completed)'
+      : '';
     console.log(
       `[${gitLabProjectId.replace(
-        "%2F",
-        "/"
+        '%2F',
+        '/'
       )} #${issueNumber}] ${new Date().toLocaleString()} ${status} ${fullCompleteMessage}`
     );
   }
@@ -139,13 +143,13 @@ export function configReadline() {
 
 export function setUpSyncHotkey(gitLabProjectId: string, issueNumber: string) {
   process.stdin.setRawMode(true);
-  process.stdin.on("keypress", async (_, key) => {
-    if (key.ctrl && key.name === "c") {
+  process.stdin.on('keypress', async (_, key) => {
+    if (key.ctrl && key.name === 'c') {
       process.exit();
-    } else if (!key.ctrl && !key.meta && !key.shift && key.name === "s") {
+    } else if (!key.ctrl && !key.meta && !key.shift && key.name === 's') {
       console.log(`You pressed the sync key`);
       syncChecklist(gitLabProjectId, issueNumber);
-    } else if (!key.ctrl && !key.meta && !key.shift && key.name === "e") {
+    } else if (!key.ctrl && !key.meta && !key.shift && key.name === 'e') {
       console.log(`You pressed the end key`);
       await syncChecklist(gitLabProjectId, issueNumber);
       const projectName = getGitLabProjectConfigById(gitLabProjectId)?.name;